+++ oc_lib/models/operation_vente_view.py
@@ -42,7 +42,7 @@ class OperationVenteView(db.Model, Repository):
     beneficiaire_pm_registre_commerce = db.Column(db.Integer, nullable=False)
     beneficiaire_pm_centre = db.Column(db.Integer, nullable=False)
     beneficiaire_pm_raison_sociale = db.Column(db.String(100))
-    beneficiaire_pm_idce = db.Column(db.Integer)
+    beneficiaire_pm_idce = db.Column(db.BigInteger)
     # Si pp dispose d'autorisation
     numero_autorisation = db.Column(db.Integer, nullable=True)
     statut = db.Column(db.Integer)
